fix: implement team review findings for precision, security, and performance
- Store monetary amounts as integer cents instead of Float to avoid
  precision issues (Transaction.amount, Budget.amount_limit,
  RecurrenceGroup.expected_amount)
- Add Pydantic field validators to convert cents↔dollars at the API boundary
- Rewrite dashboard analytics to use SQL aggregation (SUM, AVG, COUNT,
  GROUP BY) instead of loading all rows into Python memory
- Replace in-memory CategoryRule matching with a SQL LIKE query
- Fix transaction update to allow setting nullable fields to None
- Narrow CSV import exception handling with per-row logging
- Mark imported categories as is_system=False (user-created)
- Add 10 MB file size limit on CSV upload endpoint
- Sanitize CSV text values against spreadsheet formula injection
- Enable Alembic batch mode for SQLite ALTER TABLE compatibility

+++ backend/app/models/budget.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, DateTime, Float, ForeignKey, Integer, func
+from sqlalchemy import Column, DateTime, ForeignKey, Integer, func
 
 from app.models.base import Base
 
@@ -8,5 +8,5 @@ class Budget(Base):
 
     id = Column(Integer, primary_key=True, autoincrement=True)
     category_id = Column(Integer, ForeignKey("categories.id"), nullable=False, unique=True)
-    amount_limit = Column(Float, nullable=False)  # monthly budget limit
+    amount_limit = Column(Integer, nullable=False)  # cents; monthly budget limit
     created_at = Column(DateTime, server_default=func.now(), nullable=False)
